moves task functions to import.js
